requirements.txt
flask
gunicorn
crewai
langchain-google-vertexai
google-cloud-aiplatform
python-dotenv
pandas